Fix for ftps. Abstract  connection provider
